backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow==2.18.1
numpy==1.26.4
Pillow==10.1.0
